Show Windows console when isviewer is present.
