+++ os/winapi/console.c
@@ -1,0 +1,31 @@
+//
+// os/winapi/console.c
+//
+// Functions for manipulating the console.
+//
+// CEN64: Cycle-Accurate Nintendo 64 Emulator.
+// Copyright (C) 2015, Tyler J. Stachecki.
+//
+// This file is subject to the terms and conditions defined in
+// 'LICENSE', which is part of this source code package.
+//
+
+#include "cen64.h"
+#include <windows.h>
+
+// "Hides" the console window (after waiting for input).
+void hide_console(void) {
+  printf("\n");
+  system("PAUSE");
+
+  FreeConsole();
+}
+
+// "Unhides" the console window.
+void show_console(void) {
+  if (AttachConsole(ATTACH_PARENT_PROCESS) == 0)
+    AllocConsole();
+
+  freopen("CONOUT$", "wb", stdout);
+  freopen("CONOUT$", "wb", stderr);
+}
